requirements.txt
entsoe_py==0.5.10
pandas==2.1.2
python-dotenv==1.0.0
